+++ src/ansys_sphinx_theme/search/__init__.py
@@ -23,7 +23,13 @@
 
 from sphinx.application import Sphinx
 
-from ansys_sphinx_theme.search.fuse_search import create_search_index
+from ansys_sphinx_theme.search.fuse_search import (
+    ALL_NODES,
+    LITERAL,
+    PARAGRAPHS,
+    TITLES,
+    create_search_index,
+)
 
 
 def update_search_config(app: Sphinx) -> None:
@@ -36,8 +42,17 @@ def update_search_config(app: Sphinx) -> None:
     """
     theme_static_options = app.config.html_theme_options.get("static_search", {})
     theme_static_options["keys"] = ["title", "text"]
-    theme_static_options["threshold"] = theme_static_options.get("threshold", 0.5)
+    theme_static_options["threshold"] = theme_static_options.get("threshold", 0.2)
     theme_static_options["limit"] = theme_static_options.get("limit", 10)
+    app.add_config_value("index_patterns", {}, "html")
+    app.config.html_theme_options["static_search"] = theme_static_options
 
 
-__all__ = ["create_search_index", "update_search_config"]
+__all__ = [
+    "create_search_index",
+    "update_search_config",
+    "LITERAL",
+    "PARAGRAPHS",
+    "TITLES",
+    "ALL_NODES",
+]